% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.R
\name{attrs}
\alias{attrs}
\title{Attributes}
\usage{
attrs(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name value pairs giving attributes to map to
variables. The names for x and y aspects are typically omitted because
they are so common; all other aspects must be named.}
}
\description{
Attributes to draw horizon plot.
}
\section{Coordinates}{

Valid coordinates (depending on layer).
\itemize{
  \item{\code{x}, \code{y}}
  \item{\code{val}}
}
}

