% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsdata}
\alias{tsdata}
\title{Random data}
\format{
An data.frame object with 5000 rows and 3 columns
\describe{
  \item{\code{date}}{ Date sequence}
  \item{\code{value}}{ A number}
  \item{\code{group}}{ A group name}
}
}
\usage{
tsdata
}
\description{
Generated data for examples.
}
\keyword{datasets}
