% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon.R
\name{horizon}
\alias{horizon}
\title{Initialise}
\usage{
horizon(data, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A \code{data.frame} or \link[tibble]{tibble}.}

\item{...}{Attributes, specified using \code{\link{attrs}}.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Id of element.}
}
\description{
Initialise an horizon plot.
}
